can now process and determine ice quantities
It currently calculates the minimum ice quantity for the year and sums it for both north and south. An important point as that these minimums occur in opposite seasons so its not actually quite reality. Also the search for north season needs to be imporved to be from midyear to midyear not end to end so that its min for that years seaons and not half of both winter seasons in a year.